fix lint bugs
